Set enclave location for attestation test utility.
Platform verification testing had demonstrated the rather
irritating problem that the test-remote-target utility was not
finding its enclave when it was run from an installation.  The
utility was always expecting the enclave to be in the current
working directory.

This was noted to be secondary to the fact that the Makefile in
the SRDE/Attestation directory had not been updated to detect
that the ENCLAVE_DIR make variable had been set and int turn set
a compilation define in the CFLAGS directory.  This commit adds
the necessary infrastructure to properly define the location of
the enclave.