Update OpenSSL.
This commit documents the upgrade to OpenSSL 1.1.1d for both the
standard and enclave versions of the library.  This is a long
term support release of OpenSSL that will be maintained until
September 11th, 2023.

This version of OpenSSL is notable for support of TLS 1.3 as well
as Curve25519 based ECDSA signatures.

As part of the upgrade work was done to convert the Intel
provided SGX 'shim' libraries from C++ to standard C.  This
means that enclaves that are C based only do not need to be
linked with the C++ standard library (libsgx_tcxx.a).  This
result in a decrease in enclave size of approximately 330
kilobytes or approximately one third of a megabyte.

It is unclear why the shim code for a C-only library such as
OpenSSL was written in C++.  All but two of the interface files
were converted to standard C simply by removing the extern "C"
directives in the file.

OpenSSL references the strdup() function and this function is not
provided by the Intel supplied standard C library.  This had been
addressed by placing the strdup function in the C++ library.  It
is unclear why a standard C library would have been placed in the
C++ library.

In order to address this the SRDEfusion library was updated to
to implement the strdup() function in order to retain
independence from the standard C++ library.

The Makefiles that were being used to create enclaves that used
OpenSSL support were modified to drop the references to the