Implement enclave uninitialization.
In the 2.x development series Intel implemented an
'uninitialization ECALL' that is used to carry out termination
procedures for an enclave.  One of the objectives of this ECALL
is to support the atexit() functionality.

The de-initialization ECALL implemented as the -5 slot in the
ECALL dispatch table.  The 2.x series also implemented an ECALL
that is referenced in the -4 slot to create additional Task
Control Structures (TCS's).

The HurdLib/SRDEFusion library uses atexit() functionality to
track object allocations in order to detect 'lost' objects which
result in leaked memory and possible security issues from memory
that does not get cleared.

Prior to the uninitialization ECALL the 'lost' object
functionality was not available.

This commits adds a call to the uninitialization ECALL in the
SRDEenclave->whack method in order to implement the enclave
retirement and gain access to atexit() functionality.

In an implementation that is different from Intel's an OCALL
table is passed to the de-initialization in order to give the
ECALL access to the stdio output functions.  This is required
in order to allow the SRDEfusion atexit functionality to print
out indications of lost objects.

The _check_state() function in the Origin object was modified to
indicate that lost objects had been detected in the enclave.
This was done in order to differentiate messages from the enclave
from lost object messages generated by standard userspace.

The test-fusion unit test driver and its associated enclave were
modified to implement a lost object unit test.  This is not run
as part of the standard unit tests but can be specificaly
requested by the -L command-line arguement to the test-fusion
utility.

The newly implemented support for lost object detection indicated
there were issues with the test-naaaim unit test.  These
regressions were fixed so that the SRDEnaaaim standard user tests
do not generate lost object messages.