Fix lost objects in SEALEDblob object.
The newly added support for proper atexit() function handling
found a problem with the object that implements enclave sealing
of data.

The AES256_cbc object was being accidentally double initialized.
This caused lost object messages to be generated.

This commit reduces the CPP macro initialization that was causing
the problem.