Fix lost objects in SEALkey object.
The object that implements enclave sealing keys had a regression
that was causing Buffer objects to be lost.  This was addressed
by doing the initialization of the Buffer objects that hold the
key and initialization vector in the object construct.

This allows the _generate_iv_key() function to simply issue a
call to the ->reset method of the two objects rather then trying
to use a condition to detect if the objects should be initialized
or reset.