Implement ability to retrieve RSA key modulus.
This update implements the following new method for RSA key
management:

RSAkey->get_modulus

For an object that is implementing either a public, private,
or a generated key type this method will return the modulus
that is used for the asymmetric keypair.

The RSAkey_test unit test utility was modified to take the
following command-line arguements:

modulus [public|private] keyfile

Where the public and private keywords specify the type of key to
be loaded from the keyfile.  The utility will output a
Buffer->hprint representation of the modulus.

The test-naaaim unit test enclave was updated to print out the
modulus for a key that is generated inside of an enclave.