Implement framework for mode 2 PossumPipe support.
This commit provides the framework for a new authentication mode
for PossumPipe connections.  This mode uses a public key as the
identity for the connection.  This mode is designed as a
bootstrap mode to support Host Specific Enclave Authentication
(HSEA).

The changes in this commit do not actually implement any of the
needed infrastructure but instead implements a unit test that
will be used for implementing the desired functionality.

Since the new mode changes the API interfaces to the PossumPipe
object current uses of this object are updated to allow them to
compile without error.