Implement mode 2 PossumPipe support.
The framework that was implemented in the previous commit was
updated to include an initial implementation of mode 2
authentication.  This authentication mode uses a private RSA key
as the platform identity source.

The PossumPacket object was updated to add the following three
methods in order to support mode 2 authentication:

->set_schedule2

->create_packet2

->decode_packet2

All three of these methods take an RSAkey object rather then an
IDtoken object as the identity source in order to set the key
schedule.  This key schedule and identity is used by the
->create_packet2 and ->decode_packet2 methods to create and
process the mode 2 authentication packet.

The PossumPipe object was updated to add the find2_client()
function in order to locate the public key that is the
counter-party of the private key used as the host identity.

The following two functions were added to the PossumPipe object
to facilitate sending and receiving a platform attestation quote:

send_platform_report()

receive_platform_report()

Since one of the initial objectives of mode 2 authentication
support is to provision IAS credentials to a host it is not
possible for the client to issue a SPID to the communication
counter-party to generate an attestation report.  In order to
support mutual attestation the model in mode 2 authentication is
to have the server side of the connection self-generate an
attestation report for review by the client.

This can be considered secure since Intel generates a signature
over the report that can be validated by an external certificate.
The host side of the connection self-generates the attestation
report and returns the ASCII output of the REST call to the client
to validate the signature and then act on the attestation
information.

Since mode 2 attestation does not have the ability to have
enclave software status available, a synthentic status value
compiled into the PossumPipe object is used to setup the
connection.  The client uses its private key to encrypt its
connection nonce.  The host side of the connection will
use this nonce as the HMAC key to authenticate its use of the
connection status.

Additional work is needed on this infrastructure but at this
point the new code is capable of supporting the PossumPipe unit
test.