Add passphrase support for RSAkey private key loading.
The SGX port of OpenSSL does not have support for the default PEM
passphrase solicitation infrastructure.  Attempting to load a
private RSA key that is passphrase protected results in the PEM
I/O functions returning an error.

The update provides a password callback for the ->load_private
method that uses the previously developed Prompt object to
solicit a passphrase from standard userspace.

This provides a userspace experience that is consistented with
the standard userspace implementation of the RSAkey object.  That
instance will use the default solicitation mechanism if an
encrypted key is loaded via this method.