Validate attestation report signature.
The mode 2 client code was not explicitly validating the report
signature.  This commit uses the SRDEquote->validate_report
method to explicitly validate the host supplied signature using
the IAS supplied signature and certificate.

It should be noted that the SRDEquote method contains a private
copy of the Intel IAS root certificate so the validation has a
root of trust separate from the supplied report.