Implement initial framework for APIKEY attrestation reports.
There are two separate methods implemented for authentication to
the Intel Attestation Services for the purposes of obtaining an
attestation report.

The initial implementation was based on a key/certificate pair.
This required a certificate that was registered through a known
Certificate Authority.

This was followed by the implementation of a second
authentication method based on establishing a shared secret with
Intel.  This is known as the API key approach.  The API key
approach uses the following header in the REST request:

--header=Ocp-Apim-Subscription-Key: NNNNNNNNNN

Where NNNNNNNNNN is replaced with the ASCII hexadecimal encoding
of the 128 bit shared key.

These keys have a separate SPID associated with them.

An additional arguement was added to the ->generate_report method
of the SRDEquote object.  This arguement is expected to be a
String object that contains the hexadecimal encoding of the API
key that is to be used for the quote.

The enclave version of the SRDEquote object was modified to copy
the API key value into a 33 byte array in the OCALL marshalling
object.  A boolean value is also set in the object to indicate to
standard userspace that API key authentication is to be used.

This commit does not implement API authentication but implements
all of the basic changes needed for the source tree to compile
with demonstrated functionality of cert/key requests.