Implement support for APIkey attestation reports.
This update provides an initial implementation of IAS attestation
report requests that are authenticated through the APIkey API.

The test-remote-target unit test was modified to support testing
of the API.  The utility now takes a -k command-line arguement
which is used to specify a file containing the API key value that
is to be used for authenticating the report request.  The -s
command-line arguement needs to be paired with this to select the
SPID that is registered with that API key.

Infrastructure was also implemented to allow the trusted
implementation of the SRDEquote object to submit an API key to
standard userspace.  This commit does not have support in the
unit test enclave at this time but will be added in a subsequent
commit.

URL's for both the production and development versions of the
APIkey service were added to the SRDEquote enclave but currently
only the production service URL is currently used.  Support for
specifying that the development implementation should be used
will be added in a subsequent commit.