Implement support for IAS develoment services.
This commit implements support for allowing the SRDEquote object
to use the IAS development rather then production services.  When
a Trusted Computing Base (TCB) recovery is executed Intel typically
implements the changes in their development service first rather
then production.  Having access to the development service is
thus useful to verify how a software stack will function after a
TCB recover.

The ->development method was added to the SRDEquote object to
provide a mechanism for changing the object to development mode.
This sets a flag in the object state information that causes the
development services URL rather then the production services URL
to be used.