Implement enclave based APIkey report generation.
This commit implements the ability for enclave based report
generation to use the APIkey authentication mode.  Support was
also added to request reports from either the production or the
development services.

The enclave interface code for ECALL1 was substantially
overhauled to use a single structure to convey all of the call
arguements into the enclave.  This is consistent with the
approach now being taken with newer enclaves.