Fix object leak in standard userspace report generation.
The test-remote-target utility was generating messages for lost
trusted objects.

This was secondary to the untrusted code that was implementing
the report generation using ECALL 0 to generate the enclave
target report that will be submitted for generation of an
attestation report.

ECALL 0 was generarating a Curve25519 public key for the reportdata
field.  This is used to test local attestation and the object is
released when ECALL 0 is run in mode 2 to complete the local
attestation exchange.  Since mode 2 is never run by the
test-remote-target utility the object is lost.

A mode 3 was implemented to signal that ECALL 0 should call the
object destructor after the report is generated.