Update documentation for current changes.
The commit documents the fact that PossumPipe connections now
have support for mode 2 authentication which allows a private key
to be used as the extant identity for a host.  This
infrastructure is designed to support bootstrapping hosts for
Host Specific Enclave Authentication.

Also documented is that the SRDEquote object now supports using
the APIkey API for requesting IAS attestation reports.  The
APIkey API will be the preferred method of authenticatio moving
forward and is based on a shared secret key with Intel.