Implement IAS nonce support.
The Intel Authentication Service supports the ability to send up
to a 128 bit hexadecimally encoded nonce as part of an
attestation report request.  This allows a relying party to
verify that the report is 'fresh' and is not a replayed report.

This commit updates both the the standard and enclave versions of
the SRDEquote object to optionally include a NONCE value.  The
following method was added to both object instances to support
this:

->set_nonce

Using this method to set the nonce causes the ->generate_report
method to add the nonce to the IAS report request.

The ->decode_report checks to see if the report includes a
'nonce' JSON field.  If so it is parsed out and used to populate
the state specific Buffer object that holds the nonce.