Add OpenSSL implementation of AES128-CMAC.
The SGX eco-system uses a Cipher block Message Authentication
Code (CMAC) based on the AES128 symmetric block cipher throughout
its implementation.  Up until this point the SRDE was using the
implementation provided by the Intel SDK.  This update adds both
standard an enclave mode implementations of an AES128_cmac
implementation based on the OpenSSL library.

Unit tests are supplied for both the standard and enclave
implementations of the object.  The unit tests use a key and
message from the NIST standards suite and provides, in addition
to the computed MAC, the expected output.

The AES128_cmac boject will be used to eventually replace the
SRDEcmac object that was implemented to encapsulate the
functionality provided by the Intel SDK.