Add missing unit test files.
Unfortunately when the Prompt and X509cert objects were added the
unit tests for the standard versions of the objects were not
included.  Since lib/Makefile does not build the unit tests by
default this was missed until an inspect of 'git status' was
incorporated into the workflow.

This commit adds the missed unit tests.  Given that the missing
unit tests were not breaking the build it was deemed inadvisable
to spent time rewinding and rebuilding the history to incorporate
these unit tests at their inclusion time.