Replace AES-CMAC implementation with OpenSSL implementation.
This commit converts the SRDEcmac object into using the
AES128_cmac object rather then the Intel SDK implementation.  The
only consumer of this was the srde-provision utility and
functionality was confirmed after this replacement.

The SRDEcmac object will be subsequently retired.

The Changes file was updated to document that the SRDE has
converted to this new implementation.