Update local attestation unit test.
The local attestation unit test had suffered from 'bit rot'
secondary to the changes that implemented dynamic OCALL support.

In preparation for work to develop an object that allows
management of report generation and verification a general update
to the utility and its associated enclave was conducted.

Most principally the unit test was updated to implement dynamic
OCALL support.  The standard normal userspace implementations of
OCALL support for the SRDEfusion and SRDEnaaaim are now used by
the enclave.

The SRDEenclave->setup utility was used to instantiate enclave
and local open coded support for its equivalent was removed from
the test-local-target utility.

A missed object destructor for the Curve25519 object was added so
the utility is now 'clean' with respect to object cleanup.