Implement Report object.
As a preamble to the development of an object to mediate
communications between two enclaves this commit implements an
enclave version of an object that provides support for the
generation and verification of enclave reports.  The following
three methods are implemented:

->generate_report

->validate_report

->get_targetinfo

The ->get_targetinfo method implements a 'null' report that
provides a description of the enclave that executes the method.
This target infomration is used by another enclave to generate a
report that can only be verified by the enclave that generated
the target information.

The ->generate_report method generates a report that describes
the enclave that invokes the method against target information
that has been returned by the ->get_targetinfo method.

The ->validate_report method enables the original enclave that
generated the target information to verify that the report had
been generated by an enclave on the local platform.  It does so
by verifying the AES128-CMAC message authentication code
generated over the body of the report.  The key used to seed the
AES128-CMAC algorithm can only be generated by an enclave with
the characteristics specified by the ->get_targetinfo method.

The LocalSource and LocalTarget enclaves were updated to use this
infrastructure to implement the test-local-target unit test.
This update demonstrated that the previous implementation of the
unit test was not properly validating the LocalTarget enclave.
This update addresses that regression.