Implement generation of a self-report.
A common paradigm in enclave programming is to implement a
self-report that provides a description of the enclave that is
executing.  The ->get_report method was added to the Report
object to implement this functionality.

It operates simply by calling the ->get_targetinfo method
following by submission of this information to the
->generate_report method with a NULL descriptor for the
report data information.