Convert SEALkey object to use Report object.
The SEALkey uses the enclave report mechanism in order to
obtain the enclave identity parameters which the sealing key will
be generated against.  This was previously supported by open
coding of the ENCLU[EREPORT] functionality.

This commit replaces that open coded functionality with use of
the Report object.  This allows stack based allocation of
structures with significant alignment constraints to be removed.