Implement SRDEpipe framework.
This commit implements the basic framework for an object that
will be used to mediate enclave<->enclave communications based on
a security context established with local attestation.

The model implemented is similar to what has been done for the
Duct and SRDEquote objects.  The OCALL is handled by the
SRDEpipe_mgr code that maintains a set of SRDEpipe objects in
standard space, one for each SRDEpipe implemented in enclave
context.

At this point object construction and destruction is implemented
with the rest of the functionality layered on top of this
framework.