Implement setup method for SRDEpipe object.
This commit implements the ->setup method for both the standard
and enclave implementations of the SRDEpipe object.  The ->setup
method is responsible for loading and initializing the enclave
that will be the target of the communications.

In addition to the setup of the enclave the slot number that will
be used for communications with the enclave is set in the state
information of the standard implementation of the SRDEpipe
object.