Implement initial SRDEpipe connection support.
This update provides an initial implementation of the
infrastructure needed to establish a security context between two
communicating enclaves.  This security context is a preamble for
the ability to exchange general information between the two
enclaves.

A third ECALL was added to the LocalTarget enclave to serve as a
unit test for the SRDEpipe endpoint.

The following methods were added to the SRDEpipe in order to
support establishing the security context:

->bind

->connect

->accept

->connected

As noted in the sources the ->bind method is unused in the
enclave implementation of the SRDEpipe object.  It is used
by standard userspace to execute the ECALL's into the target
enclave by the SRDEpipe manager implementation.

An enclave initiating a connection to a target enclave uses the
->connect method to do so.  When the method returns successfully
a security context is available to support further
communications.

The ->accept method is used by the remote method to respond to a
->connect request.  Secondary to the need to use a didactic
exchange in order to setup the security context the ->accept
method must be invoked twice by the target enclave.

In order to support the need to invoke the ->accept method
multiple times the ->connected method is implemented to provide
an accessor method for determining the connection status.  The
strategy is to use a NULL value for the SRDEpipe object to
indicate that the first stage ->accept should be called.  A
second invocation of the target ECALL uses the ->connected method
to determine whether or not the second invocation of the ->accept
method should be made.

At this stage of development the LocalSource/LocalTarget SRDEpipe
unit test is issuing lost object messages at the end of the unit
test.  This will be subsequently addressed by a protocol call
which will indicate the connection is to be closed.