Stop report data memory leak.
When generating an attestation quote the Report object was using
the size of the reportdata array when copying from the Buffer
object that was supplying the report data, rather then the
reported size of the Buffer object.  In the case of using a
Curve25519 public key this resulted in leaking 32 bytes of data
from the heap.

This was addressed by first unconditionally zeroing the contents
of the aligned on stack array used to pass the report data to the
ENCLU[EREPORT] instruction.  A conditional check was added to
make sure that the size of the Buffer object containing the
report data does not exceed the size of the on stack array.

With these protections in place the size of the Buffer object
containing the report data is used to copy the data into the
report data array.  This stems the leakage of heap information
into the report data.