Derive encryption key and initialization vector.
Best practice is to not use the computed shared secret from a
Curve25519 elliptic key exchange directory.  Instead the
recommendation is to hash the secret before using as an
encryption key.

This update hashes the shared secret to generate the encryption
key.

The SRDEpipe conduit uses the same strategy as PossumPipe where
both sides of the communications conduit generate an encryption
vector by hashing a common starting point.  After initial hashing
to generate the encryption key the key is hashed an additional
100 times to yield the starting point for the initialization
vector.

This strategy means that if the initialization vector were to be
come known multiple first pre-image resistance challenges would
need to be solved in order to determine the value of the
encryption key.