Implement connection verification.
The SRDEpipe implementation did not have positive implementation
of acknowledgement of the completion of the connection.  This
update implements a method for securely indicating that the
target enclave has completed the connection.

The accepting end of the connection generates a SHA256_HMAC
checksum of the initialization vector using the encryption key as
the HMAC password.  The result MAC is used as the reportdata in
an enclave report generated against the target enclave.  The
report is returned to the enclave executing the connection.

The initiating enclave verifies the integrity of the report and
then computes the SHA245_HMAC checksum using its copy of the
encryption key and initialization vector.  This value is
compared to the version returned in the reportdata element of
the enclave report.

Each side of the security context rehashes the initialization
vector after its use in this role.