Implement SRDEpipe packet transmission and reception.
This update adds support for the following two methods to the
SRDEpipe object:

->send_packet

->receive_packet

Which implement the ability for the initiating enclave to send an
encrypted and integrity protected packet to the target enclave.

The architecture for packet transmission and reception was
largely taken from the PossumPipe object.  Unlike the PossumPipe
object this code does not use per packet authentication keys.
Future updates will focus on making both SRDEpipe and PossumPipe
use a similar architecture.

The encrypted packet contains an ASN1 encoded set of data that
includes a packet type and a potential data packet for the remote
enclave.