Implement secured SRDEpipe closing.
With support for sending packets that have confidentiality and
integrity protections in place this update implements the ability
to send a pipe close event to the remote target.  The following
method was added to both the standard and enclave versions of the
SRDEpipe object in order to implement this:

->close

The OCALL to execute the destructor for the standard
implementation of the object was moved from the ->whack method to
the ->close method so that only the initiator side of the pipe
would execute the OCALL.

Since this protocol command is executed in the context of the
security channel an aggressor cannot forge a close request.

At this point in time there is one lost object from standard
userspace.  This will be addressed in the next commit.