Move SRDEpipe implementation to ->setup method.
A hard to locate object loss was found to be secondary to the
fact that the enclave implementation of the SRDEpipe object was
initializing the standard userspace implementation in the object
constructor.

In the current implementation there is no need for the target
SRDEpipe instance to have an implementation in standard
userspace.  In order to avoid initializing a userpspace instance
of the object by the object in the target enclave the userspace
implementation was moved to the ->setup object which is only
called by the initiating instance of the object.

With this change in place the unit test reports no lost objects
in either standard or enclave space.