Implement SRDEpipe packet return.
This commit provides the final infrastructure needed to make the
SRDEpipe object fully usable, the ability to return a packet from
the target enclave to the initiating enclave.

One of the major pieces of infratructure required for this is a
pool of Buffer objects with one Buffer object associated with
each standard userspace SRDEpipe implementation.  This Buffer
object provides a data pool for holding data returned from the
target enclave and directed to the initiating enclave.

The userspace implementation of the SRDEpipe object had support
added to it to review the status of the ECALL that conveyed the
packet into the target enclave.   The target enclave indicates in
the return how much data is needed to hold the return packet.
The Buffer object is resized to meet this requirement and a
second ECALL initiated to return the buffer.

The Buffer object does not currently have the ability to surface
the limit size of the object, only the size that is in use.  A
subsequent enhancement would be to service the limit size since
this may limit the need for the second ECALL.  Given that the
SRDEpipe object is currently not performance sensitive this issue
will be tabled for now.