Gate SRDEpipe debugging information.
This update implements a per object state variable that gates
whether or not debugging messages are output by the SRDEpipe
object.  This value is universally set to off by the object
initialization code.

No method is provided to enable debugging at this time and the
value is left as a compilation controllable option.  This is to
avoid the potential for coercing an object into divulging
security sensitive information.

Minor cosmetic changes were made to the unit test target enclave.