Update changes for Report and SRDEpipe object.
This commit updates the file that documents changes to indicate
the implementation of the Report and SRDEpipe objects.

The Report object encapsulates the functionality needed for
generating local attestation reports.  It also implements support
for generating an enclave self-report that is a commonly used
paradigm for an enclave to determine its identity
characteristics.

The SRDEpipe builds on the Report object to encapsulate all of
the functionality needed to support secured communiations between
two local enclaves.  This functionality is similar to the
PossumPipe object and will be updated to be more similar in
subsequent updates.