Initial implementation of attestation provider infrastructure.
This commit provides an initial framework for the remote
attestation provider architecture.  This architecture is intended
to allow any platform to implement its own remote attestation
service provider support.

The following two enclaves were implemented to provide the
framework for this service:

Attestation.signed.so

Provisioner.signed.so

The Attestation enclave will ultimately implement the
functionality needed to generate an attestation report from Intel
Athentication Services (IAS) for any enclave on the platform.  An
SRDEpipe object will be used to connect to the Attestation
enclave to return the report in a secure fashion.

The Provisioner enclave implements the credential service
provider which will run on a provisioning server that is intended
to vend credentials for IAS connectivity to Attestation enclaves
that can authenticate to it.  Connections are authenticated via
mode 2 PossumPipe connections from hosts that have a public key
registered with the provisioning service.

In addition to these two enclaves the following two utilities
have an initial implementation:

srde-provisioner

srde-attestation

The srde-provisioner utility is the deamon application that
drives execution of the Provisioner enclave on a credential
provisioning server.

The srde-attestation utility will be a general purpose tool for
managing SRDE based remote attestation on a platform.  Currently
its primary role is to drive the Attestation enclave in
provisioning mode where it initiates a connection to the
provisioning service for downloading and deploying credentials
for the platform instance of an Attestation enclave.

This commit is designed to be a staging commit.  At the current
time it supports the implementation of a mode 2 PossumPipe
connection to an instance of the Provisioner service.