Implement enclave File object.
This update provides a minimal implementation of a File object
for applications running in enclave context.  At the current time
the following methods are implemented:

->open_ro

->open_rw

->read_Buffer

->slurp

->write_Buffer

->whack

This provides the basic functionality needed for an enclave to
create and write contents of a file and to read the contents of a
file.

The methodology implemented is similar to what has been done for
other objects that manipulate resources in standard userspace on
behalf of an object in enclave context.  The SRDEfile_mgr
function was implemented to maintain a cohort of File and Buffer
objects that implement the standard userspace methods on behalf
of the same methods in the enclave version of the File object.

The test-fusion unit test was updated to implement a unit test
for the File object.  This unit test consists of opening a file
and writing contents to it.  The File object is destroyed resulting
in the standard userspace implementation of the file being
closed.  The file is opened in read-only mode and the contents
read with the ->slurp method and displayed.

Additional File methods will be implemented as needed.