Address lost object in SEALkey object.
Incidental testing with the naaaim-test utility indicated the
utility was indicating a significant number of objects lost from
enclave space.

A review of the object indicated that when the object was
modified to use the Report object, a destructor call was not
added to the _generate_iv_key() function.  This update adds that
destructor and the test-naaaim utility now runs without lost
objects.