Restore functionality to enclave File->write_buffer operations.
It was noted in collateral testing of the attestation tools that
the File->write_Buffer method when invoked from enclave was not
generating any output to the file that had been opened.  An
investigation of the problem indicated that somehow the code for
the ->slurp method had been substituted for the ->write_Buffer
code in the srdefile_write_Buffer() function.

This commit returns the proper code to the
srdefile_write_Buffer() function that restored the ability of the
object to send data to the output file.