Fix information exchange issues with SRDEpipe.
Functional use of the SRDEpipe object indicated there were
functional regressions in its implementation.  This arose from
not testing the case where the enlcave was NOT requesting an
extension of the Buffer object that data was being returned to.

One of the major issues that was missing was a clause to handle
the copy of data back in the case where there was no need to
expand the Buffer object.

This suggested that using the Buffer supplied to the
->send_packet method of the standard userspace implementation of
the SRDEpipe object.  In order to properly address this issue a
local Buffer was instantiated in the ->send_packet method to
handle conveyance of data into and out of the enclave.

Changes were also required to the interface function for the
SRDEpipe ECALL.  The incoming buffer was not being cloned,
instead the buffer contents itself was being used.  The interface
function was modified to clone the buffer so the enclave works
on a protected version of a data buffer which is then copied back
into locally instantiated Buffer memory.

In a related issue it was noted that the I/O Buffer object that
was co-allocated with the userspace implementation of the
SRDEpipe was not being released.  A call to the destructor for
that object was added to the SRDEpipe manager function.