Fix no packet return from SRDEpipe ECALL.
Further testing indicated a regression where the remote target of
an SRDEpipe endpoint was not able to specify that no packet was
being returned after a packet had been processed.  This was
traced to the fact that although the return bufr was being reset
to zero the bufr_size element of the OCALL structure was not
being reset to zero if no data packet was being returned.

This commit universally sets the bufr return size to zero and
then allows the value to be reset if the return buffer has a
non-zero content size.