Properly initialize SRDEquote enclave object.
Functional experience with implementation of the Attestation
services enclave indicated that the enclave version of the
SRDEquote object was improperly selecting the use of IAS
development services.  This was traced to the incomplete
initialization of the state information in the enclave version of
the object.

This commit fully and explicitly initializes all of the state
information in the object which addressed the functional
regression.