Implement remote attestation services infrastructure.
This rather major commit provides an initial implementation of a
centralized facility for generating remote attestation reports.
The generation and return of a remote attestation report is
implemented in the following enclave:

Attestation.signed.so

This enclave is designed to be the only entity on the system that
has access to the SPID and APIkey needed to authenticate to Intel
Attestation Services (IAS).  The model is for a client to open an
SRDEpipe connection with the Attestation enclave and negotiate
for the generation of an enclave report and subsequent generation
of a quote and attestation report for that enclave.

In order to securely deploy credentials to the platform the
following enclave and related infratructure is impelemented:

Provisioner.signed.so

This enclave is designed to listen for mode 2 PossumPipe
connections from an Attestation enclave.  After authenticating
the connection the Provisioner deploys a SPID and APIkey to the
initiator enclave.  That enclave encrypts and caches a copy of
the authentication credentials for subsequent use.

The following utility daemon was implemented to support running
and management of the Provisioner enclave:

srde-provisioner

The following utility program was implement to drive the
provisioning request by the Attestation enclave:

srde-attestation

This utility currently runs in two modes.  Provisioning is
requested via the -P command-line switch while the -T
command-line switch is used to request a sample attestation and
report for the LocalTarget.signed.so enclave.

The objective moving forward will be to encapsulate all of the
functionality into an Attestation enclave given the need to
handle state negotiations for the SRDEpipe object.