Functional cleanups for credential provisioning.
Since the PossumPipe object has not been updated to use the
enclave attestation service it requires that a SPID be available.
Initial testing had used a hard coded PID but in order to
simplify operation the Provisioning enclave was modified to read
a SPID from the following canonical location:

/opt/IDfusion/etc/spid.txt

Sine the provisioner is presumably on a secured host standard
discretionary acccess controls can be applied to that file to
limit access to it until the PossumPipe object can be modified to
use the attestation service enclave.

The attestation enclave was modified to test for whether or not
the ENCLAVE_DIR CPP define was set.  If the macro is set the
acpv.idfusion.net hostname is used for the provisioning server,
if not localhost is used.  This is designed to allow the
localhost to be used for testing and a production build will
configure the enclave to use the canonical DNS name.

The CPP define that declares the pathname to the attestation
credentials was used to define the file that the credentials will
be read from.  This replaces a file in the local directory that
had been used during initial development.